added components and vars css
