+++ src/components/AccordionSet.vue
@@ -1,0 +1,13 @@
+<template>
+  <h1>test</h1>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style>
+
+</style>
