+++ src/components/SectionDescription.vue
@@ -1,0 +1,11 @@
+<template>
+  <span>FAQ's</span>
+  <h1>Frequently Asked Questions</h1>
+  <p>It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. </p>
+</template>
+
+<script>
+export default {
+
+}
+</script>
