+++ src/styles/global/global.sass
@@ -3,10 +3,13 @@
   padding: 0
   box-sizing: border-box
 
+body 
+  font-family: $font-primary
+
 .container 
   max-width: 1084px
   margin: 0 auto
-  padding: 0 15px
+  padding: 15vh 15px
 
 .section-row 
   display: flex
