+++ src/styles/global/vars.sass
@@ -1,0 +1,9 @@
+$font-primary: sans-serif
+$font-secondary: sans-serif
+
+$color-primary: red
+$color-secondary: red
+
+$--h1-set: 21px
+$--h2-set: 20px
+$--p-set: 16px
