+++ src/styles/main.css
@@ -4,10 +4,14 @@
   box-sizing: border-box;
 }
 
+body {
+  font-family: sans-serif;
+}
+
 .container {
   max-width: 1084px;
   margin: 0 auto;
-  padding: 0 15px;
+  padding: 15vh 15px;
 }
 
 .section-row {
