+++ src/styles/main.sass
@@ -1,1 +1,2 @@
+@import './global/vars'
 @import './global/global'
